src/datasets/requirements/base.txt
django
djangorestframework
django_neomodel
neo4j
